+++ spec/cuprum/operation_spec.rb
@@ -106,4 +106,72 @@ RSpec.describe Cuprum::Operation do
       end
     end
   end
+
+  describe '#to_proc' do
+    let(:proc)   { operation.to_proc }
+    let(:result) { proc.call }
+    let(:expected_error) do
+      Cuprum::Errors::CommandNotImplemented.new(command: operation)
+    end
+
+    it { expect(operation).to respond_to(:to_proc).with(0).arguments }
+
+    it { expect(operation.to_proc).to be_a Proc }
+
+    it { expect(operation.to_proc).to be proc }
+
+    it { expect(result).to be operation }
+
+    it { expect(result.success?).to be false }
+
+    it { expect(result.error).to be == expected_error }
+
+    it { expect(nil.yield_self(&operation)).to be operation }
+
+    it 'should call the operation with arguments' do
+      # rubocop:disable RSpec/SubjectStub
+      arguments = %w[ichi ni san]
+
+      allow(operation).to receive(:call)
+
+      proc.call(*arguments)
+
+      expect(operation).to have_received(:call).with(*arguments)
+      # rubocop:enable RSpec/SubjectStub
+    end
+
+    it 'should call the operation with keywords' do
+      # rubocop:disable RSpec/SubjectStub
+      keywords = { ichi: 1, ni: 2, san: 3 }
+
+      allow(operation).to receive(:call)
+
+      proc.call(**keywords)
+
+      expect(operation).to have_received(:call).with(**keywords)
+      # rubocop:enable RSpec/SubjectStub
+    end
+
+    context 'when the implementation is defined' do
+      let(:described_class) do
+        process = implementation
+
+        Class.new(super()) do |klass|
+          klass.send(:define_method, :process, &process)
+        end
+      end
+      let(:implementation) { ->(str) { str.upcase } }
+      let(:value)          { 'Greetings, programs!' }
+      let(:result)         { proc.call(value) }
+      let(:expected_value) { value.upcase }
+
+      it { expect(result).to be operation }
+
+      it { expect(result.success?).to be true }
+
+      it { expect(result.value).to be == expected_value }
+
+      it { expect(value.yield_self(&operation).value).to be == expected_value }
+    end
+  end
 end