+++ spec/cuprum/error_spec.rb
@@ -3,9 +3,10 @@
 require 'cuprum/error'
 
 RSpec.describe Cuprum::Error do
-  subject(:error) { described_class.new(message: message) }
+  subject(:error) { described_class.new(message: message, **properties) }
 
-  let(:message) { nil }
+  let(:message)    { nil }
+  let(:properties) { {} }
 
   describe '::new' do
     it 'should define the constructor' do
@@ -13,6 +14,7 @@ RSpec.describe Cuprum::Error do
         .to be_constructible
         .with(0).arguments
         .and_keywords(:message)
+        .and_any_keywords
     end
   end
 
@@ -31,14 +33,20 @@ RSpec.describe Cuprum::Error do
       it { expect(error == Object.new.freeze).to be false }
     end
 
-    describe 'with an Error with no message' do
+    describe 'with an error with no message' do
       let(:other) { described_class.new }
 
       it { expect(error == other).to be true }
     end
 
-    describe 'with an Error with non-matching message' do
+    describe 'with an error with non-matching message' do
       let(:other) { described_class.new(message: 'An error occurred.') }
+
+      it { expect(error == other).to be false }
+    end
+
+    describe 'with an error with non-matching properties' do
+      let(:other) { described_class.new(color: 'red') }
 
       it { expect(error == other).to be false }
     end
@@ -55,6 +63,14 @@ RSpec.describe Cuprum::Error do
       include_context 'when there is an error subclass'
 
       let(:other) { Spec::Error.new(message: 'An error occurred.') }
+
+      it { expect(error == other).to be false }
+    end
+
+    describe 'with an Error subclass with non-matching properties' do
+      include_context 'when there is an error subclass'
+
+      let(:other) { Spec::Error.new(color: 'red') }
 
       it { expect(error == other).to be false }
     end
@@ -113,6 +129,175 @@ RSpec.describe Cuprum::Error do
         let(:other) { Spec::Error.new(message: message) }
 
         it { expect(error == other).to be false }
+      end
+    end
+
+    context 'when initialized with properties' do
+      let(:properties) { { color: 'red', shape: 'möbius strip' } }
+
+      describe 'with nil' do
+        # rubocop:disable Style/NilComparison
+        it { expect(error == nil).to be false }
+        # rubocop:enable Style/NilComparison
+      end
+
+      describe 'with an Object' do
+        it { expect(error == Object.new.freeze).to be false }
+      end
+
+      describe 'with an Error with non-matching message' do
+        let(:other) { described_class.new(message: 'An error occurred.') }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with no properties' do
+        let(:other) { described_class.new }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with non-matching properties' do
+        let(:other) { described_class.new(color: 'blue', shape: 'torus') }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with partially-matching properties' do
+        let(:other) { described_class.new(color: 'red', shape: 'torus') }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with matching properties' do
+        let(:other) { described_class.new(color: 'red', shape: 'möbius strip') }
+
+        it { expect(error == other).to be true }
+      end
+    end
+
+    context 'when initialized with properties and a message' do
+      let(:message)    { 'Something went wrong.' }
+      let(:properties) { { color: 'red', shape: 'möbius strip' } }
+
+      describe 'with nil' do
+        # rubocop:disable Style/NilComparison
+        it { expect(error == nil).to be false }
+        # rubocop:enable Style/NilComparison
+      end
+
+      describe 'with an Object' do
+        it { expect(error == Object.new.freeze).to be false }
+      end
+
+      describe 'with an Error with no message or properties' do
+        let(:other) { described_class.new }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with non-matching message' do
+        let(:other) { described_class.new(message: 'An error occurred.') }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with matching message' do
+        let(:other) { described_class.new(message: message) }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with non-matching properties' do
+        let(:other) { described_class.new(color: 'blue', shape: 'torus') }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with partially-matching properties' do
+        let(:other) { described_class.new(color: 'red', shape: 'torus') }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with matching properties' do
+        let(:other) { described_class.new(color: 'red', shape: 'möbius strip') }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an Error with matching properties and message' do
+        let(:other) do
+          described_class.new(
+            message: message,
+            color:   'red',
+            shape:   'möbius strip'
+          )
+        end
+
+        it { expect(error == other).to be true }
+      end
+    end
+
+    describe 'when the Error subclass defines custom comparable properties' do
+      include_context 'when there is an error subclass'
+
+      subject(:error) { Spec::Error.new(message: message, **properties) }
+
+      before(:example) do
+        Spec::Error.define_method(:color) do
+          @comparable_properties[:color] # rubocop:disable RSpec/InstanceVariable
+        end
+
+        Spec::Error.define_method(:comparable_properties) do
+          { color: color }
+        end
+      end
+
+      describe 'with an error with no message or properties' do
+        let(:other) { Spec::Error.new }
+
+        it { expect(error == other).to be true }
+      end
+
+      describe 'with an error with non-matching message' do
+        let(:other) { Spec::Error.new(message: 'An error occurred.') }
+
+        it { expect(error == other).to be true }
+      end
+
+      describe 'with an error with non-matching properties' do
+        let(:other) { Spec::Error.new(color: 'blue') }
+
+        it { expect(error == other).to be false }
+      end
+
+      describe 'with an error with matching properties' do
+        let(:other) { Spec::Error.new(shape: 'torus') }
+
+        it { expect(error == other).to be true }
+      end
+
+      context 'when initialized with properties' do
+        let(:properties) { { color: 'red', shape: 'möbius strip' } }
+
+        describe 'with an error with no properties' do
+          let(:other) { Spec::Error.new }
+
+          it { expect(error == other).to be false }
+        end
+
+        describe 'with an error with non-matching properties' do
+          let(:other) { Spec::Error.new(color: 'blue', shape: 'möbius strip') }
+
+          it { expect(error == other).to be false }
+        end
+
+        describe 'with an error with matching properties' do
+          let(:other) { Spec::Error.new(color: 'red', shape: 'torus') }
+
+          it { expect(error == other).to be true }
+        end
       end
     end
     # rubocop:enable RSpec/NestedGroups